backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.2
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
google-cloud-aiplatform>=1.35.0
google-auth>=2.23.0
pandas>=2.1.0
